requirements.txt
aws-cdk-lib==2.139.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.56
pandas>=2.2.1
numpy>=1.26.4
cdk-nag>=2.28.55
dynamo-pandas>=1.3.0
h5py>=3.11.0
Cython>=3.0.10
kaleido==0.2.1; sys_platform == 'linux'
kaleido==0.1.0post1; sys_platform == 'win32'
pre-commit==4.0.1
